requirements.txt
Flask==3.0.0
PyPDF2==3.0.0
python-docx==0.8.11
Werkzeug==3.0.0
gunicorn==20.1.0
google-genai>=1.0.0
